WIP experimenting with making calls fallible
It's not even clear I want to change these things; in some cases,
possibly we can prove that we would have already hit the recursion
limit beforehand or similar.

Also, these changes should be split up so as to touch as little as
possible, since it's pretty easy to end up with a single commit that
touches way too much code at once, which is impossible to review.

Mostly, I'm going to keep expanding this commit until it catches the
test case I added, and then go back and carefully examine some of
the (potentially dangerous) assumptions I've made here.