Plumb environment-id to sqllogictest
Currently, sqllogictest will leave clusterd processes around when it
either crashes or fails -- in the latter case, this is because it
process::exit()s on failure, which causes the Drop
impl (KillOnDropChild) in the process orchestrator to not fire.

I considered putting a separate environment-id for sqllogictest in
/tmp, but I think if we're going this route instead of killing all of
sqllogictest's descendants, we might as well share the environment
with environmentd.